requirements.txt
gradio
torch
torchvision
transformers
Pillow
pytest
requests
timm
numpy<2
